Refactor duplicate output code
