+++ classes/output/assess_by_month.php
@@ -34,6 +34,8 @@ use local_assessfreq\frequency;
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class assess_by_month {
+    use generate_assess_by_month_chart;
+
     /**
      * Generate the markup for the process summary chart,
      * used in the smart media dashboard.
@@ -46,44 +48,7 @@ class assess_by_month {
         // Get events for the supplied year.
         $frequency = new frequency();
         $yeardata = $frequency->get_events_due_by_month($year);
-        $seriesdata = [];
         $charttitle = get_string('assessbymonth', 'local_assessfreq');
-        $result = [];
-
-        // There is always 12 months in a year,
-        // even if we don't have data for them all.
-        for ($i = 1; $i <= 12; $i++) {
-            if (!empty($yeardata[$i])) {
-                $seriesdata[] = $yeardata[$i]->count;
-            } else {
-                $seriesdata[] = 0;
-            }
-        }
-
-        // Create chart object.
-        $events = new \core\chart_series($charttitle, $seriesdata);
-        $labels = [
-            get_string('jan', 'local_assessfreq'),
-            get_string('feb', 'local_assessfreq'),
-            get_string('mar', 'local_assessfreq'),
-            get_string('apr', 'local_assessfreq'),
-            get_string('may', 'local_assessfreq'),
-            get_string('jun', 'local_assessfreq'),
-            get_string('jul', 'local_assessfreq'),
-            get_string('aug', 'local_assessfreq'),
-            get_string('sep', 'local_assessfreq'),
-            get_string('oct', 'local_assessfreq'),
-            get_string('nov', 'local_assessfreq'),
-            get_string('dec', 'local_assessfreq'),
-        ];
-
-        $chart = new \core\chart_bar();
-        $chart->add_series($events);
-        $chart->set_labels($labels);
-
-        $result['hasdata'] = true;
-        $result['chart'] = $chart;
-
-        return $result;
+        return $this->generate_chart($yeardata, $charttitle);
     }
 }
